requirements.txt
line-bot-sdk
Flask
Werkzeug
opencc-python-reimplemented
beautifulsoup4
youtube-transcript-api
pyyaml
psycopg2
gunicorn